barcode_hw/requirements.txt
qrcode~=8.1
python-barcode~=0.15.1
pillow~=11.1.0